+++ docs/Changelog.rst
@@ -9,7 +9,9 @@ Changelog
 **Bug fixes**
 
   * Add more logging information to snapshot delete method #582 (untergeek)
-  
+  * Improve default timeout, logging, and exception handling for `seal` command
+    #583 (untergeek)
+
 3.5.0 (16 March 2016)
 ---------------------